// src/api/management/management.did.d.ts
export interface canister_status_result {
  'memory_metrics' : {
    'wasm_binary_size' : bigint,
    'wasm_chunk_store_size' : bigint,
    'canister_history_size' : bigint,
    'stable_memory_size' : bigint,
    'snapshots_size' : bigint,
    'wasm_memory_size' : bigint,
    'global_memory_size' : bigint,
    'custom_sections_size' : bigint,
  },
  'status' : { 'stopped' : null } |
    { 'stopping' : null } |
    { 'running' : null },
  'memory_size' : bigint,
  'cycles' : bigint,
  'settings' : definite_canister_settings,
  'query_stats' : {
    'response_payload_bytes_total' : bigint,
    'num_instructions_total' : bigint,
    'num_calls_total' : bigint,
    'request_payload_bytes_total' : bigint,
  },
  'idle_cycles_burned_per_day' : bigint,
  'module_hash' : [] | [Uint8Array | number[]],
  'reserved_cycles' : bigint,
}
// ...
export interface definite_canister_settings {
  'freezing_threshold' : bigint,
  'wasm_memory_threshold' : bigint,
  'controllers' : Array<Principal>,
  'reserved_cycles_limit' : bigint,
  'log_visibility' : log_visibility,
  'wasm_memory_limit' : bigint,
  'memory_allocation' : bigint,
  'compute_allocation' : bigint,
}
// ...
export type log_visibility = { 'controllers' : null } |
  { 'public' : null } |
  { 'allowed_viewers' : Array<Principal> };